refactor(parser): modularize and clean parser components
• Extracted multiple functions from `parser_parse.c` into dedicated files:
  - `end_command.c`
  - `ft_lstget.c`
  - `parser_special_redirect_in.c` (refactored with helper)
  - `parser_special_redirect_out.c`
  - `parser_special_redirect_append.c`
  - `parser_special_redirect_heredoc.c`
  - `get_redirect_token.c`
• Updated `Makefile` to include new source files under the parser section.
• Added function prototypes to `parser.h` for better modularity and maintainability.
• Simplified `parser_handle_special` and `parser_step` logic for readability.
